Refactor to make XYModem modular using proxy classes
