+++ XYModem/examples/example_xmodem.cpp
@@ -27,49 +27,50 @@ int main (int argc, [[maybe_unused]] char* argv[])
                              readTimeoutMultiplier,
                              writeTimeoutConstant,
                              writeTimeoutMultiplier);
-    SerialHandler serialHandler (serialDevice);
 
-    XModem xmodem (serialHandler, logType (), 2);
-    ghc::filesystem::path file;
+    auto serialHandler = std::make_shared<SerialHandler>(serialDevice);
+    auto logger = std::make_shared<Spdlogger>();
+    XModemSender xmodem (serialHandler, logger);
+
+    std::string filePath;
     bool withHex = false;
-    std::string lastParam;
 
     cliParser.setCommands({
-        {"-h", [](std::string_view){
-           spdlog::info("This program allows you to send one file "
+        {"-h", [&logger](std::string_view){
+           logger->info("This program allows you to send one file "
                          "using the XModem file transfer protocol.\n Options "
-                         ":\n    -h : get help \n    -f \"filepath1\"  -id "
-                         "\"VID/PID of the device\""
+                         ":\n    -h : get help \n    -f \"filepath1\"  -port "
+                         "\"serial port of the device\""
                          "\"filepath2\" ... : send one or more files \n");}},
-        {"-f", [&file](std::string_view value){
-                spdlog::info (value);
-                file=value.data();
+        {"-f", [&filePath, &logger](std::string_view value){
+                logger->info (value);
+                filePath=value.data();
         }},
         {"--hex", [&withHex](std::string_view){withHex = true;}},
-        {"-port", [&serialDevice](std::string_view val){
+        {"-port", [&serialDevice, &logger](std::string_view val){
               const auto availablePorts = serial::list_ports ();
               const auto serialPort =
                   std::find_if (availablePorts.begin (),
                                 availablePorts.end (),
-                                [&val] (auto& device) {
+                                [&val, &logger] (auto& device) {
                                     if(val.empty())
                                     {
-                                        spdlog::info ("Available devices");
-                                        spdlog::info (device.description);
-                                        spdlog::info (device.port);
-                                        spdlog::info (device.hardware_id);
+                                        logger->info ("Available devices");
+                                        logger->info (device.description);
+                                        logger->info (device.port);
+                                        logger->info (device.hardware_id);
                                     }
                                     return device.port == val;
                                 });
               if (serialPort != availablePorts.end ())
               {
-                  spdlog::info (serialPort->port);
+                  logger->info (serialPort->port);
                   serialDevice.setPort (serialPort->port);
                   serialDevice.open ();
               }
               else
               {
-                  spdlog::error ("Unable to find the requested device");
+                  logger->error ("Unable to find the requested device");
               }
         }}
     });
@@ -78,32 +79,32 @@ int main (int argc, [[maybe_unused]] char* argv[])
     {
         // parse command line arguments and launch related commands
         cliParser.parse (argc, argv);
-        if (!file.empty ())
+        if (!filePath.empty ())
         {
             xmodem.transmit (
-                file, [] (float) {}, [] () { return false; }, withHex);
+                std::make_shared<xymodem::DesktopFile>(filePath), [] (float) {}, [] () { return false; }, withHex);
         }
     }
     catch (serial::IOException& e)
     {
-        spdlog::warn (e.what ());
+        logger->warn (e.what ());
         assert (false);
     }
     catch (XYModemExceptions::Timeout const& e)
     {
-        spdlog::warn (e.what ());
+        logger->warn (e.what ());
     }
     catch (XYModemExceptions::CouldNotOpenFile const& e)
     {
-        spdlog::warn (e.what ());
+        logger->warn (e.what ());
     }
     catch (XYModemExceptions::TransmissionAborted const& e)
     {
-        spdlog::warn (e.what ());
+        logger->warn (e.what ());
     }
     catch (std::exception const& e)
     {
-        spdlog::warn (e.what ());
+        logger->warn (e.what ());
         assert (false);
     }
 }
