+++ XYModem/include/XYModem/XYModemConst.h
@@ -1,10 +1,37 @@
 #pragma once
+#include <string>
 #include <array>
-#include <tuple>
-#include <functional>
-#include <string_view>
-#include "FileTransferProtocol.h"
+#include <chrono>
 using namespace std::literals::chrono_literals;
+
+namespace XYModemExceptions
+{
+struct Timeout : public std::exception
+{
+    long long timeoutSeconds = 0;
+
+    Timeout (long long timeoutSeconds) : timeoutSeconds (timeoutSeconds){};
+
+    virtual const char* what () const throw () { return "Timeout"; }
+};
+struct CouldNotOpenFile : public std::exception
+{
+    std::string absolutePath;
+    std::string response;
+
+    CouldNotOpenFile (std::string absolutePath) : absolutePath (absolutePath){};
+
+    virtual const char* what () const throw () { return "Could not open file"; }
+};
+
+struct TransmissionAborted : public std::exception
+{
+    virtual const char* what () const throw ()
+    {
+        return "Transmission aborted.";
+    }
+};
+} // namespace XYModemExceptions
 
 using int64 = long long;                                                                                            
 namespace xyModemConst
