+++ XYModem/tests/TestXModemTransitions.cpp
@@ -8,151 +8,151 @@ namespace xymodem
 class XModemTest : public ::testing::Test
 {
 public:
-    XModemTest () : serialHandler (serialDevice), xModem (serialHandler) {}
+    XModemTest () : deviceHandler (std::make_shared<SerialHandler>(serialDevice)), xModem (deviceHandler) {}
 
     serial::Serial serialDevice;
-    SerialHandler serialHandler;
-    XModem xModem;
+    std::shared_ptr<SerialHandler> deviceHandler;
+    XModemSender xModem;
 };
 
 TEST_F (XModemTest, TestBeginXModem)
 {
     EXPECT_EQ (xModem.getNextState (xyModemConst::C,
-                                    XModem::waitingStart,
-                                    XModem::undefined,
-                                    XModem::stateTransitions),
-               XModem::sendingPacket);
+                                    XModemSender::waitingStart,
+                                    XModemSender::undefined,
+                                    XModemSender::stateTransitions),
+               XModemSender::sendingPacket);
 }
 
 TEST_F (XModemTest, TestSendNextPacket)
 {
-    xModem.guards.set (XModem::packetsLeft, 2);
+    xModem.guards.set (XModemSender::packetsLeft, 2);
     EXPECT_EQ (xModem.getNextState (xyModemConst::ACK,
-                                    XModem::sendingPacket,
-                                    XModem::undefined,
-                                    XModem::stateTransitions),
-               XModem::sendingPacket);
+                                    XModemSender::sendingPacket,
+                                    XModemSender::undefined,
+                                    XModemSender::stateTransitions),
+               XModemSender::sendingPacket);
 }
 
 TEST_F (XModemTest, TestSendNextPacketButNoPacketsLeft)
 {
-    xModem.guards.set (XModem::packetsLeft, 0);
+    xModem.guards.set (XModemSender::packetsLeft, 0);
     EXPECT_EQ (xModem.getNextState (xyModemConst::ACK,
-                                    XModem::sendingPacket,
-                                    XModem::undefined,
-                                    XModem::stateTransitions),
-               XModem::endOfTransmission);
+                                    XModemSender::sendingPacket,
+                                    XModemSender::undefined,
+                                    XModemSender::stateTransitions),
+               XModemSender::endOfTransmission);
 }
 
 TEST_F (XModemTest, TestRetryingButReachedRetriesLimit)
 {
-    xModem.guards.set (XModem::retries, 11);
+    xModem.guards.set (XModemSender::retries, 11);
     EXPECT_EQ (xModem.getNextState (xyModemConst::NAK,
-                                    XModem::retryingPacket,
-                                    XModem::undefined,
-                                    XModem::stateTransitions),
-               XModem::abort);
+                                    XModemSender::retryingPacket,
+                                    XModemSender::undefined,
+                                    XModemSender::stateTransitions),
+               XModemSender::abort);
 }
 
 TEST_F (XModemTest, TestRetryingPacket)
 {
-    xModem.guards.set (XModem::retries, 0);
+    xModem.guards.set (XModemSender::retries, 0);
     EXPECT_EQ (xModem.getNextState (xyModemConst::NAK,
-                                    XModem::sendingPacket,
-                                    XModem::undefined,
-                                    XModem::stateTransitions),
-               XModem::retryingPacket);
+                                    XModemSender::sendingPacket,
+                                    XModemSender::undefined,
+                                    XModemSender::stateTransitions),
+               XModemSender::retryingPacket);
 }
 TEST_F (XModemTest, TestReretryingPacket)
 {
-    xModem.guards.set (XModem::retries, 5);
+    xModem.guards.set (XModemSender::retries, 5);
     EXPECT_EQ (xModem.getNextState (xyModemConst::NAK,
-                                    XModem::retryingPacket,
-                                    XModem::undefined,
-                                    XModem::stateTransitions),
-               XModem::retryingPacket);
+                                    XModemSender::retryingPacket,
+                                    XModemSender::undefined,
+                                    XModemSender::stateTransitions),
+               XModemSender::retryingPacket);
 }
 
 TEST_F (XModemTest, TestEOTButNAKReceived)
 {
-    xModem.guards.set (XModem::retries, 0);
+    xModem.guards.set (XModemSender::retries, 0);
     EXPECT_EQ (xModem.getNextState (xyModemConst::NAK,
-                                    XModem::endOfTransmission,
-                                    XModem::undefined,
-                                    XModem::stateTransitions),
-               XModem::retryingEOT);
+                                    XModemSender::endOfTransmission,
+                                    XModemSender::undefined,
+                                    XModemSender::stateTransitions),
+               XModemSender::retryingEOT);
 }
 
 TEST_F (XModemTest, TestEOTButNAKReceivedAndNoRetriesLeft)
 {
-    xModem.guards.set (XModem::retries, 11);
+    xModem.guards.set (XModemSender::retries, 11);
     EXPECT_EQ (xModem.getNextState (xyModemConst::NAK,
-                                    XModem::retryingEOT,
-                                    XModem::undefined,
-                                    XModem::stateTransitions),
-               XModem::abort);
+                                    XModemSender::retryingEOT,
+                                    XModemSender::undefined,
+                                    XModemSender::stateTransitions),
+               XModemSender::abort);
 }
 
 TEST_F (XModemTest, TestUnknownCharacterReceived)
 {
     EXPECT_EQ (xModem.getNextState (23,
-                                    XModem::sendingPacket,
-                                    XModem::undefined,
-                                    XModem::stateTransitions),
-               XModem::undefined);
+                                    XModemSender::sendingPacket,
+                                    XModemSender::undefined,
+                                    XModemSender::stateTransitions),
+               XModemSender::undefined);
 }
 
 TEST_F (XModemTest, TestTransmissionFinished)
 {
     EXPECT_EQ (xModem.getNextState (xyModemConst::ACK,
-                                    XModem::endOfTransmission,
-                                    XModem::undefined,
-                                    XModem::stateTransitions),
-               XModem::transmissionFinished);
+                                    XModemSender::endOfTransmission,
+                                    XModemSender::undefined,
+                                    XModemSender::stateTransitions),
+               XModemSender::transmissionFinished);
 }
 
 TEST_F (XModemTest, TestTransmissionFinishedAfterEOTRetry)
 {
     EXPECT_EQ (xModem.getNextState (xyModemConst::ACK,
-                                    XModem::retryingEOT,
-                                    XModem::undefined,
-                                    XModem::stateTransitions),
-               XModem::transmissionFinished);
+                                    XModemSender::retryingEOT,
+                                    XModemSender::undefined,
+                                    XModemSender::stateTransitions),
+               XModemSender::transmissionFinished);
 }
 
 TEST_F (XModemTest, TestSendingPacketButCAN)
 {
     EXPECT_EQ (xModem.getNextState (xyModemConst::CAN,
-                                    XModem::sendingPacket,
-                                    XModem::undefined,
-                                    XModem::stateTransitions),
-               XModem::abort);
+                                    XModemSender::sendingPacket,
+                                    XModemSender::undefined,
+                                    XModemSender::stateTransitions),
+               XModemSender::abort);
 }
 
 TEST_F (XModemTest, TestSendEOTButCAN)
 {
     EXPECT_EQ (xModem.getNextState (xyModemConst::CAN,
-                                    XModem::endOfTransmission,
-                                    XModem::undefined,
-                                    XModem::stateTransitions),
-               XModem::abort);
+                                    XModemSender::endOfTransmission,
+                                    XModemSender::undefined,
+                                    XModemSender::stateTransitions),
+               XModemSender::abort);
 }
 
 TEST_F (XModemTest, TestRetryingPacketButCAN)
 {
     EXPECT_EQ (xModem.getNextState (xyModemConst::CAN,
-                                    XModem::retryingPacket,
-                                    XModem::undefined,
-                                    XModem::stateTransitions),
-               XModem::abort);
+                                    XModemSender::retryingPacket,
+                                    XModemSender::undefined,
+                                    XModemSender::stateTransitions),
+               XModemSender::abort);
 }
 
 TEST_F (XModemTest, TestRetryingEOTButCAN)
 {
     EXPECT_EQ (xModem.getNextState (xyModemConst::CAN,
-                                    XModem::retryingEOT,
-                                    XModem::undefined,
-                                    XModem::stateTransitions),
-               XModem::abort);
+                                    XModemSender::retryingEOT,
+                                    XModemSender::undefined,
+                                    XModemSender::stateTransitions),
+               XModemSender::abort);
 }
 } // namespace xymodem
