+++ XYModem/tests/TestYModemTransitions.cpp
@@ -8,107 +8,107 @@ namespace xymodem
 class YModemTest : public testing::Test
 {
 public:
-    YModemTest () : serialHandler (serialDevice), yModem (serialHandler){};
+    YModemTest () : deviceHandler (std::make_shared<SerialHandler>(serialDevice)), yModem (deviceHandler){};
 
     serial::Serial serialDevice;
-    SerialHandler serialHandler;
-    YModem yModem;
+    std::shared_ptr<SerialHandler> deviceHandler;
+    YModemSender yModem;
 };
 
 TEST_F (YModemTest, TestBeginYModem)
 {
     EXPECT_EQ (yModem.getNextState (xyModemConst::C,
-                                    YModem::waitingStart,
-                                    YModem::undefined,
-                                    YModem::stateTransitions),
-               YModem::sendingHeader);
+                                    YModemSender::waitingStart,
+                                    YModemSender::undefined,
+                                    YModemSender::stateTransitions),
+               YModemSender::sendingHeader);
 }
 
 TEST_F (YModemTest, TestBeginXModemTransmission)
 {
     EXPECT_EQ (yModem.getNextState (xyModemConst::ACK,
-                                    YModem::sendingHeader,
-                                    YModem::undefined,
-                                    YModem::stateTransitions),
-               YModem::xModemTransmission);
+                                    YModemSender::sendingHeader,
+                                    YModemSender::undefined,
+                                    YModemSender::stateTransitions),
+               YModemSender::xModemTransmission);
 }
 
 TEST_F (YModemTest, TestRetryingHeader)
 {
     EXPECT_EQ (yModem.getNextState (xyModemConst::NAK,
-                                    YModem::sendingHeader,
-                                    YModem::undefined,
-                                    YModem::stateTransitions),
-               YModem::retryingHeader);
+                                    YModemSender::sendingHeader,
+                                    YModemSender::undefined,
+                                    YModemSender::stateTransitions),
+               YModemSender::retryingHeader);
 }
 
 TEST_F (YModemTest, TestReretryingHeader)
 {
-    yModem.guards.set (YModem::retries, 2);
+    yModem.guards.set (YModemSender::retries, 2);
     EXPECT_EQ (yModem.getNextState (xyModemConst::NAK,
-                                    YModem::retryingHeader,
-                                    YModem::undefined,
-                                    YModem::stateTransitions),
-               YModem::retryingHeader);
+                                    YModemSender::retryingHeader,
+                                    YModemSender::undefined,
+                                    YModemSender::stateTransitions),
+               YModemSender::retryingHeader);
 }
 
 TEST_F (YModemTest, TestTooManyRetries)
 {
-    yModem.guards.set (YModem::retries, 11);
+    yModem.guards.set (YModemSender::retries, 11);
     EXPECT_EQ (yModem.getNextState (xyModemConst::NAK,
-                                    YModem::retryingHeader,
-                                    YModem::undefined,
-                                    YModem::stateTransitions),
-               YModem::abort);
+                                    YModemSender::retryingHeader,
+                                    YModemSender::undefined,
+                                    YModemSender::stateTransitions),
+               YModemSender::abort);
 }
 
 TEST_F (YModemTest, TestUnknownCharacterReceived)
 {
     EXPECT_EQ (yModem.getNextState (23,
-                                    YModem::sendingHeader,
-                                    YModem::undefined,
-                                    YModem::stateTransitions),
-               YModem::undefined);
+                                    YModemSender::sendingHeader,
+                                    YModemSender::undefined,
+                                    YModemSender::stateTransitions),
+               YModemSender::undefined);
     EXPECT_EQ (yModem.getNextState (23,
-                                    YModem::retryingHeader,
-                                    YModem::undefined,
-                                    YModem::stateTransitions),
-               YModem::undefined);
+                                    YModemSender::retryingHeader,
+                                    YModemSender::undefined,
+                                    YModemSender::stateTransitions),
+               YModemSender::undefined);
 }
 
 TEST_F (YModemTest, TestWaitingStartButCAN)
 {
     EXPECT_EQ (yModem.getNextState (xyModemConst::CAN,
-                                    YModem::waitingStart,
-                                    YModem::undefined,
-                                    YModem::stateTransitions),
-               YModem::abort);
+                                    YModemSender::waitingStart,
+                                    YModemSender::undefined,
+                                    YModemSender::stateTransitions),
+               YModemSender::abort);
 }
 
 TEST_F (YModemTest, TestSendingHeaderButCAN)
 {
     EXPECT_EQ (yModem.getNextState (xyModemConst::CAN,
-                                    YModem::sendingHeader,
-                                    YModem::undefined,
-                                    YModem::stateTransitions),
-               YModem::abort);
+                                    YModemSender::sendingHeader,
+                                    YModemSender::undefined,
+                                    YModemSender::stateTransitions),
+               YModemSender::abort);
 }
 
 TEST_F (YModemTest, TestRetryingHeaderButCAN)
 {
     EXPECT_EQ (yModem.getNextState (xyModemConst::CAN,
-                                    YModem::sendingHeader,
-                                    YModem::undefined,
-                                    YModem::stateTransitions),
-               YModem::abort);
+                                    YModemSender::sendingHeader,
+                                    YModemSender::undefined,
+                                    YModemSender::stateTransitions),
+               YModemSender::abort);
 }
 
 TEST_F (YModemTest, TestXModemTransmissionButCAN)
 {
     EXPECT_EQ (yModem.getNextState (xyModemConst::CAN,
-                                    YModem::xModemTransmission,
-                                    YModem::undefined,
-                                    YModem::stateTransitions),
-               YModem::abort);
+                                    YModemSender::xModemTransmission,
+                                    YModemSender::undefined,
+                                    YModemSender::stateTransitions),
+               YModemSender::abort);
 }
 } // namespace xymodem
